// src/types/dropwave/Products.ts
export interface Product {
  id: number;
  name: string;
  slug: string;
  tag_id: number | null;
  maincategory_id: number;
  subcategory_id: number | null;
  brand_id: number | null;
  label_id: number | null;
  tax_id: number | null;
  tax_status: string;
  shipping_id: number | null;
  preview_type: string;
  preview_video: string | null;
  preview_content: string | null;
  trending: number;
  status: number;
  video_url: string | null;
  track_stock: number;
  stock_order_status: string;
  price: number;
  sale_price: number;
  product_type: string | null;
  product_stock: number;
  low_stock_threshold: number;
  downloadable_product: boolean | null;
  product_weight: number | null;
  cover_image_path: string;
  cover_image_url: string;
  stock_status: string;
  variant_product: number;
  attribute_id: string;
  product_attribute: string;  
  custom_field_status: number;
  custom_field: string | null;
  description: string | null;
  detail: string | null;
  specification: string | null;
  average_rating: number;
  theme_id: string;
  store_id: number;
  created_by: number;
  is_active: boolean | null;
  created_at: string;  
  updated_at: string;  
}